registration screen, login screen new update
Renamed the logo.png in pubspec.yaml file. Created a user_model.dart and it is the model class that will be used to manage the user data. In the login_screen.dart file coded the login function.  This function will validate the email and password. If the email and password are valid, it will sign in the user. And changed the validator function for each text form fields.

In registration_screen.dart file coded the registration function which creates a new user with the email and password provided. If the user is created successfully, it posts the details to the firestore. For that the function (  postDetailsToFirestore()), coded to post the details to the firestore. This function is called when the user is created successfully.